pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "exmailer/__init__.py"

[project]
name = "exmailer"
version = "1.0.2"
description = "Send emails via Microsoft Exchange Server with HTML template support"
readme = "README.md"
authors = [{ name = "Sadegh Yazdani", email = "sadegh.yazdani@example.com" }]
license = { text = "GPL-3.0-or-later" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "exchange",
    "email",
    "microsoft",
    "outlook",
    "persian",
    "rtl",
    "notification",
    "smtp",
]
requires-python = ">=3.11"
dependencies = ["exchangelib>=5.0.0", "urllib3>=2.0.0"]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dotenv = ["python-dotenv>=1.2.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-console-scripts>=1.4.0",
    "isort>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-console-scripts>=1.4.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-autorefs>=1.0",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-section-index>=0.3",
    "pymdown-extensions>=10.0",
]

[project.urls]
Homepage = "https://github.com/aerosadegh/exmailer"
Repository = "https://github.com/aerosadegh/exmailer"
Documentation = "https://exmailer.readthedocs.io"
"Bug Tracker" = "https://github.com/aerosadegh/exmailer/issues"


[project.scripts]
exmailer = "exmailer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["exmailer"]

[tool.uv]
dev-dependencies = [
    "isort>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "pytest-cov>=4.0.0",
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.0",
]

[tool.uv.sources]


[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # Security checks
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = [
    "S101", # assert usages
    "S105", # Hardcoded passwords in test fixtures (safe placeholders)
    "S106", # Hardcoded secrets in test fixtures
    "S107", # Hardcoded password args in test fixtures
]


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "exchangelib.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = ["pytest-cov"]
minversion = "7.2"
addopts = [
    "--verbose",
    "--cov=exmailer",
    "--cov-report=term-missing",
    # "--cov-report=html",
    # "--cov-report=xml",
]

[tool.coverage.run]
source = ["exchange_emailer"]
omit = ["*/tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
